Moves model indexing/searching code into Concerns
