+++ app/models/screening_list/dpl.rb
@@ -1,6 +1,6 @@
 module ScreeningList
   class Dpl
-    extend ::Indexable
+    include Indexable
     include ScreeningList::Mappable
     self.source = {
       full_name: 'Denied Persons List (DPL) - Bureau of Industry and Security',
