+++ app/models/screening_list/dtc.rb
@@ -1,6 +1,6 @@
 module ScreeningList
   class Dtc
-    extend ::Indexable
+    include Indexable
     include ScreeningList::Mappable
     self.source = {
       full_name: 'ITAR Debarred (DTC) - State Department',
