+++ app/models/screening_list/el.rb
@@ -1,6 +1,6 @@
 module ScreeningList
   class El
-    extend ::Indexable
+    include Indexable
     include ScreeningList::Mappable
     self.source = {
       full_name: 'Entity List (EL) - Bureau of Industry and Security',
