+++ app/models/screening_list/fse.rb
@@ -1,6 +1,6 @@
 module ScreeningList
   class Fse
-    extend ::Indexable
+    include Indexable
     include ScreeningList::Mappable
     self.source = {
       full_name: 'Foreign Sanctions Evaders (FSE) - Treasury Department',
