+++ app/models/screening_list/isa.rb
@@ -1,6 +1,6 @@
 module ScreeningList
   class Isa
-    extend ::Indexable
+    include Indexable
     include ScreeningList::Mappable
     self.source = {
       full_name: 'Non-SDN Iranian Sanctions Act List (NS-ISA) - Treasury Department',
