+++ app/models/screening_list/part561.rb
@@ -1,6 +1,6 @@
 module ScreeningList
   class Part561
-    extend ::Indexable
+    include Indexable
     include ScreeningList::Mappable
     self.source = {
       full_name: 'Part 561 List (561) - Treasury Department',
