+++ app/models/screening_list/plc.rb
@@ -1,6 +1,6 @@
 module ScreeningList
   class Plc
-    extend ::Indexable
+    include Indexable
     include ScreeningList::Mappable
     self.source = {
       full_name: 'Palestinian Legislative Council List (PLC) - Treasury Department',
