+++ app/models/screening_list/uvl.rb
@@ -1,6 +1,6 @@
 module ScreeningList
   class Uvl
-    extend ::Indexable
+    include Indexable
     include ScreeningList::Mappable
     self.source = {
       full_name: 'Unverified List (UVL) - Bureau of Industry and Security',
