+++ app/models/tariff_rate/chile.rb
@@ -1,6 +1,6 @@
 module TariffRate
   class Chile
-    extend ::Indexable
+    include Indexable
     include TariffRate::Mappable
     self.source = {
       code: 'CL',
