+++ app/models/tariff_rate/el_salvador.rb
@@ -1,6 +1,6 @@
 module TariffRate
   class ElSalvador
-    extend ::Indexable
+    include Indexable
     include TariffRate::Mappable
     self.source = {
       code: 'SV',
