+++ app/models/tariff_rate/honduras.rb
@@ -1,6 +1,6 @@
 module TariffRate
   class Honduras
-    extend ::Indexable
+    include Indexable
     include TariffRate::Mappable
     self.source = {
       code: 'HN',
