+++ app/models/tariff_rate/nicaragua.rb
@@ -1,6 +1,6 @@
 module TariffRate
   class Nicaragua
-    extend ::Indexable
+    include Indexable
     include TariffRate::Mappable
     self.source = {
       code: 'NI',
