+++ app/models/tariff_rate/singapore.rb
@@ -1,6 +1,6 @@
 module TariffRate
   class Singapore
-    extend ::Indexable
+    include Indexable
     include TariffRate::Mappable
     self.source = {
       code: 'SG',
