+++ app/models/trade_article.rb
@@ -1,5 +1,5 @@
 class TradeArticle
-  extend Indexable
+  include Indexable
 
   self.settings = {}
   self.mappings = {
