+++ app/models/trade_event/dl.rb
@@ -1,6 +1,6 @@
 module TradeEvent
   class Dl
-    extend ::Indexable
+    include Indexable
     include TradeEvent::Mappable
     self.source = {
       code: 'DL',
