+++ app/models/trade_lead/canada.rb
@@ -1,6 +1,6 @@
 module TradeLead
   class Canada
-    extend ::Indexable
+    include Indexable
     include TradeLead::Mappable
     self.source = {
       code: 'CANADA',
